fix(instagram): share download folder preference helpers
